move functions to dbaccessors
